split out tabs into separate component
