+++ src/pages/workspaces/workspace/WorkspaceAttributeNotice.ts
@@ -12,7 +12,7 @@ interface WorkspaceAttributeNoticeProperties {
   workspaceRegionConstraintMessage?: string;
 }
 
-const WorkspaceAttributeNotice = (props: WorkspaceAttributeNoticeProperties) => {
+export const WorkspaceAttributeNotice = (props: WorkspaceAttributeNoticeProperties) => {
   const isReadOnly = !canWrite(props.accessLevel);
 
   return div({}, [
